Update CMake configuration and remove deprecated textlab assets
- Added linking of BULLET_LIBRARIES to the common target in CMakeLists.txt for improved physics support.
- Temporarily removed MP3 codec from client sources due to build issues, ensuring a smoother compilation process.
- Disabled dead code elimination for the Vulkan renderer to preserve essential symbols, enhancing compatibility.
- Deleted obsolete textlab configuration files, fonts, and assets to clean up the project structure and reduce clutter.
- Introduced new files for UI and keycode management, enhancing the user interface framework.

+++ src/client/mp3/towave.c
@@ -175,7 +175,7 @@ id3v1_1 *gpTAG;
 																									\
 	/* account for trailing ID3 tag in _iBytesRemaining */											\
 	gpTAG = (id3v1_1*) (((byte *)_pvData + _iBytesRemaining)-sizeof(id3v1_1));	/* sizeof = 128	*/	\
-	if (!strncmp(gpTAG->id, "TAG", 3))																\
+	if (!strncmp(gpTAG->id, "TAG", (size_t)3))																\
 	{																								\
 		_iBytesRemaining -= sizeof(id3v1_1);														\
 	}
